Modify TerminateCurrentProcess, Blacklist Checking
Description : terminate process found by PID.

Parameter :
( in ) DWORD32 targetPID : target terminated process' ID
if mode is 1, it must be NULL.
( in/out )TCHAR imagePath :
if mode is 0, it'll be getting fullimagepath of target process.
else if mode is 1, it must have list of blacklist processes' imagepath
( in ) DWORD length : number of row of the imagepath array.
if mode is 0, it must be NULL.
( in ) DWORD mode
0 -> Use PID
1 -> Auto Kill
else -> Error
Return value :
0 = Error
1 = Success